Rename message-history-manager
To chat-context-toolkit, to better represent what it contains.
