+++ assistants/codespace-assistant/pyproject.toml
@@ -9,6 +9,7 @@ dependencies = [
     "assistant-drive>=0.1.0",
     "assistant-extensions[attachments, mcp]>=0.1.0",
     "mcp-extensions[openai]>=0.1.0",
+    "chat-context-toolkit>=0.1.0",
     "content-safety>=0.1.0",
     "deepmerge>=2.0",
     "openai>=1.61.0",
@@ -29,6 +30,7 @@ assistant-extensions = { path = "../../libraries/python/assistant-extensions", e
 mcp-extensions = { path = "../../libraries/python/mcp-extensions", editable = true }
 content-safety = { path = "../../libraries/python/content-safety/", editable = true }
 openai-client = { path = "../../libraries/python/openai-client", editable = true }
+chat-context-toolkit = { path = "../../libraries/python/chat-context-toolkit", editable = true }
 
 [build-system]
 requires = ["hatchling"]
