+++ libraries/python/assistant-extensions/assistant_extensions/message_history/__init__.py
@@ -1,5 +1,5 @@
-from ._history import message_history_manager_message_provider_for
+from ._history import chat_context_toolkit_message_provider_for
 
 __all__ = [
-    "message_history_manager_message_provider_for",
+    "chat_context_toolkit_message_provider_for",
 ]
