+++ libraries/python/chat-context-toolkit/chat_context_toolkit/archive/ideas.md
