+++ libraries/python/chat-context-toolkit/chat_context_toolkit/history/_types.py
