+++ libraries/python/chat-context-toolkit/chat_context_toolkit/history/tool_abbreviations/_tool_abbreviations.py
