+++ libraries/python/chat-context-toolkit/chat_context_toolkit/virtual_filesystem/types.py
