+++ libraries/python/chat-context-toolkit/test/history/test_abbreviate_messages.py
@@ -1,9 +1,9 @@
 import sys
 from uuid import uuid4
 
-from message_history_manager.history import HistoryMessageProtocol, OpenAIHistoryMessageParam
-from message_history_manager.history._budget import abbreviate_messages
-from message_history_manager.history._types import BudgetDecision, MessageCollection, TokenCounts
+from chat_context_toolkit.history import HistoryMessageProtocol, OpenAIHistoryMessageParam
+from chat_context_toolkit.history._budget import abbreviate_messages
+from chat_context_toolkit.history._types import BudgetDecision, MessageCollection, TokenCounts
 from openai.types.chat import (
     ChatCompletionAssistantMessageParam,
     ChatCompletionMessageParam,
